pyproject.toml
[tool.poetry]
name = "wineapp"
version = "0.1.0"
description = ""
authors = ["Jun Liu <antyser@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.36.0"

ipython = "^8.26.0"
python-dotenv = "^1.0.1"
google-search-results = "^2.4.2"
apify-client = "^1.7.1"
mypy = "^1.11.0"
black = "^24.4.2"
httpx = {extras = ["http2"], version = "^0.27.0"}
bs4 = "^0.0.2"
python-slugify = {extras = ["unidecode"], version = "^8.0.4"}
markdownify = "^0.13.1"
ruff = "^0.5.5"
pre-commit = "^3.7.1"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
loguru = "^0.7.2"
modal = "^0.63.90"
langchain = "^0.2.11"
langchain-openai = "^0.1.17"
langchain-community = "^0.2.10"
langgraph = "^0.1.11"
langchain-experimental = "^0.0.63"
langserve = {extras = ["all"], version = "^0.2.2"}
selenium = "^4.23.1"
webdriver-manager = "^4.0.2"


[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.20240724"
types-python-slugify = "^8.0.2.20240310"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88  # Match Black's default line length

[tool.ruff.lint]
select = ["E", "F", "I"]  # Enable Pyflakes, pycodestyle, and isort rules
ignore = ["E501"]  # Ignore line length violations (let Black handle it)

[tool.black]
line-length = 88

[mypy]
ignore_missing_imports = true
